requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.119.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.5
llama-index-cli==0.5.3
llama-index-core==0.14.5
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.5
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-qdrant==0.8.6
llama-index-workflows==2.8.2
llama-parse==0.6.54
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
openai==1.109.1
packaging==25.0
pandas==2.2.3
pillow==12.0.0
platformdirs==4.5.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
PyMuPDF==1.26.5
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.9.18
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wrapt==1.17.3
yarl==1.22.0
